[#11] navbar/homepage-styling-fixes
Changed the styling based on screen width for crypto-card-container

Corrected the sidebar buttons location when minimized at a certain screen width so they are not cut-off. Also changed 'Coinverse' to H5 so that when minimized it was not word-wrapped.

Show more button was redirecting using a non-capitalized letter so the nav bar didn't properly show that the page was active, so I switched it.