[#22] crypto-bookmark/navbar-support
Created a new component for Bookmarks with the intention of adding it to the main Navigation Menu. To achieve this it was necessary to change the way the menu was put together--to be compartmentalized.

A new button has been added to each crypto's detail page that will be golden when that page is bookmarked or white when it is not. The navigation menu will indicate this change by showing the crypto's symbol and icon.

Scroll-bar changes from [#17] and [#18] happened to go missing with a commit but is still held in my branch and will be added back.